requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
python-jose
passlib[bcrypt]
python-dotenv
pypdf
docx2txt
alembic
requests
qdrant-client
langchain-openai
langchain-text-splitters
markdown